+++ lib/presentation/core/theme/theme_extension.dart
@@ -1,0 +1,7 @@
+import 'package:flutter/material.dart';
+
+extension ThemeExtension on BuildContext {
+  ColorScheme get myColorScheme => Theme.of(this).colorScheme;
+
+  TextTheme get myTextTheme => Theme.of(this).textTheme;
+}