+++ lib/presentation/features/home/screens/home_screen.dart
@@ -6,6 +6,7 @@ import '../../../../domain/entities/user.dart';
 import '../../../../l10n/lang_extensions.dart';
 import '../../../../l10n/language.dart';
 import '../../../components/common_error_loading_widget.dart';
+import '../../../core/theme/theme_extension.dart';
 import '../../translation/translation_bloc/translation_bloc.dart';
 import '../blocs/home_bloc/home_bloc.dart';
 
@@ -93,9 +94,7 @@ class _HomeScreenState extends State<HomeScreen> {
                                   return ListTile(
                                     title: Text(
                                       user[index].firstName ?? "",
-                                      style: Theme.of(context)
-                                          .textTheme
-                                          .headlineSmall,
+                                      style: context.myTextTheme.headlineSmall,
                                     ),
                                   );
                                 });