+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 # Changelog
 
+## 28/11/23
+- Added extension for text theme and color scheme.
+
 ## 21/11/23
 - Added docs in Readme.md